+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.45.1-dev.2](https://github.com/crimera/piko/compare/v1.45.1-dev.1...v1.45.1-dev.2) (2024-11-25)
+
+### Refactors
+
+* **Twitter:** Added toggle for `Show sensitive media` patch ([cac4fe7](https://github.com/crimera/piko/commit/cac4fe747640e37dc9155c63f23082c44963c05d))
+
+## [1.45.1-dev.1](https://github.com/crimera/piko/compare/v1.45.0...v1.45.1-dev.1) (2024-11-24)
+
+### Bug Fixes
+
+* **Twitter:** Change patch name from `Enable app icon settings` -> `Enable app icons` ([5a7ef46](https://github.com/crimera/piko/commit/5a7ef469557e5cf4ecc9ce8465d73c20501dc524))
+
 ## [1.45.0](https://github.com/crimera/piko/compare/v1.44.0...v1.45.0) (2024-11-14)
 
 ### Bug Fixes